pyproject.toml
[project]
name = "AtlasPrint-plugin"
authors = [
    { name = "3liz", email = "info@3liz.com" }
]
version = "3.4.4"
requires-python = ">=3.9"
license = "GPL-2.0-only"
license-files = ["LICENSE"]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: GIS",
]

[dependency-groups]
dev = [
    "coverage[toml]",
    {include-group = "tests"},
    {include-group = "lint"},
]
tests = [
    "pytest >= 5.3",
    "pytest-qgis",
    "semver",
    "lxml",
]
lint = [
    "ruff",
    "mypy",
    "bandit",
    "lxml-stubs",
]
packaging = [
    "qgis-plugin-package-ci; python_full_version >= '3.12'",
]

[tool.uv.sources]
qgis-plugin-package-ci = { git = "https://github.com/3liz/qgis-plugin-package-ci.git" }
pytest-qgis = { git = "https://github.com/3liz/pytest-qgis.git" }

# Ruff configuration
# See https://doc.astral.sh/ruff/configuration

[tool.ruff]
line-length = 110
target-version = "py39"
exclude = [
    ".venv",
    ".local",
    ".test/.local",
    ".cache",
]

[tool.ruff.format]
indent-style = "space"

[tool.ruff.lint]
extend-select = [
    "E", "F", "ANN", "W", "T", "COM", "RUF",
    "C4", "SIM", "TC", "RET",
]
ignore = [
    "ANN002",
    "ANN003",
    "ANN401",
    "COM812", # Conflict with format
    "SIM108", # Use ternary operator
    "RUF067", # module should only contain docstrings and re-exports
    # XXX Fix 
    "E501", # Line too long
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "T201",
    "E501",
    "ANN001",
]

[tool.ruff.lint.isort]
#lines-between-types = 1
known-third-party = [
    "qgis",
]

[tool.ruff.lint.flake8-annotations]
ignore-fully-untyped = true
suppress-none-returning = true
suppress-dummy-args = true

[tool.package-ci]
plugin_source = "atlasprint"
github_organization_slug = "3liz"
project_slug = "qgis-atlasprint"


#
# Mypy
#
[tool.mypy]
python_version = "3.9"

[[tool.mypy.overrides]]
module = [
   "qgis.*",
   "osgeo.*",
   "PIL.*",
]
ignore_missing_imports = true

#
# Coverage
#

[tool.coverage.run]
source = ["atlasprint"]
branch = true
relative_files = true

[tool.coverage.report]
precision = 2
exclude_lines = [
 "if TYPE_CHECKING",
 '\(Protocol\):$',
]
